code refactoring and fixes
